Upgrade Wagyu to version 0.5.0

+++ mapbox/geometry/wagyu/edge.hpp
@@ -39,8 +39,7 @@ struct edge {
     }
 
     template <typename T2>
-    edge(mapbox::geometry::point<T2> const& current,
-         mapbox::geometry::point<T2> const& next_pt) noexcept
+    edge(mapbox::geometry::point<T2> const& current, mapbox::geometry::point<T2> const& next_pt) noexcept
         : bot(static_cast<T>(current.x), static_cast<T>(current.y)),
           top(static_cast<T>(current.x), static_cast<T>(current.y)),
           dx(0.0) {
@@ -69,8 +68,8 @@ using edge_list_itr = typename edge_list<T>::iterator;
 
 template <typename T>
 bool slopes_equal(edge<T> const& e1, edge<T> const& e2) {
-    return (e1.top.y - e1.bot.y) * (e2.top.x - e2.bot.x) ==
-           (e1.top.x - e1.bot.x) * (e2.top.y - e2.bot.y);
+    return static_cast<std::int64_t>(e1.top.y - e1.bot.y) * static_cast<std::int64_t>(e2.top.x - e2.bot.x) ==
+           static_cast<std::int64_t>(e1.top.x - e1.bot.x) * static_cast<std::int64_t>(e2.top.y - e2.bot.y);
 }
 
 template <typename T>
@@ -83,16 +82,14 @@ inline double get_current_x(edge<T> const& edge, const T current_y) {
     if (current_y == edge.top.y) {
         return static_cast<double>(edge.top.x);
     } else {
-        return static_cast<double>(edge.bot.x) +
-               edge.dx * static_cast<double>(current_y - edge.bot.y);
+        return static_cast<double>(edge.bot.x) + edge.dx * static_cast<double>(current_y - edge.bot.y);
     }
 }
 
 #ifdef DEBUG
 
 template <class charT, class traits, typename T>
-inline std::basic_ostream<charT, traits>& operator<<(std::basic_ostream<charT, traits>& out,
-                                                     const edge<T>& e) {
+inline std::basic_ostream<charT, traits>& operator<<(std::basic_ostream<charT, traits>& out, const edge<T>& e) {
     out << "    Edge: " << std::endl;
     out << "        bot x: " << e.bot.x << " y: " << e.bot.y << std::endl;
     out << "        top x: " << e.top.x << " y: " << e.top.y << std::endl;
@@ -118,6 +115,6 @@ inline std::basic_ostream<charT, traits>& operator<<(std::basic_ostream<charT, t
 }
 
 #endif
-}
-}
-}
+} // namespace wagyu
+} // namespace geometry
+} // namespace mapbox
